Update and rename eightpuzzle.py to 8_puzzle.py
